hardware/requirements_hardware.txt
ultralytics==8.2.50
opencv-python
numpy
torch>=2.0.0
torchvision>=0.15.0
adafruit-circuitpython-mlx90640
smbus2
Pillow
matplotlib
